+++ R/global.R
@@ -2,9 +2,11 @@
 utils::globalVariables(c(
   ".", "Freq", "Freq.x", "Freq.y", "Package", "Repository", "V1", "Var1",
   "Version", "accept_cats", "auto_accept", "auto_pass", "bugs_status",
-  "decision", "decision_reason", "dep_failed", "dep_freq", "depends", "deps",
+  "decision", "decision_reason", "decision_reason_note", "dep_failed",
+  "dep_freq", "depends", "deps",
   "derived_col", "downloads_1yr", "ends_with", "exception_risk_category",
-  "final_decision", "final_risk", "final_risk_cat", "final_risk_catid",
+  "final_decision", "final_decision_reason_note", "final_risk",
+  "final_risk_cat", "final_risk_catid",
   "final_risk_id", "has_source_control", "has_vignettes", "has_website",
   "met_dec_id", "metric", "metric_type", "news_curr", "news_current",
   "package", "primary_risk_category", "repo_name", "reverse_dependencies",